easyocr/requirements.txt
flask
Pillow
easyocr
pdf2image